Fix Vulkan manifest library paths for macOS bundle (#7)
This removes the relative path prefixes `./` from the library paths in
the Vulkan manifest files, because the search paths are different when
just using filenames. It is needed for a recent change in
https://github.com/PixelGuys/Cubyz/pull/2141 that creates an application
bundle, so the Vulkan loader has to search in
`(bundle)/Contents/Frameworks` instead of where the manifest files are
located.

The `file_replace` tool is now added in two different function, maybe it
would be better to extract this.