Move miniaudio and stb_vorbis dependencies to build.zig.zon (#10)
Progress towards https://github.com/PixelGuys/Cubyz/issues/2346  

The stb_vorbis lib is included with the official miniaudio repo, so I
include it from there.

Also, miniaudio internally uses stb_vorbis for decoding ogg vorbis
files, but the stb_vorbis header must be defined first to enable that,
so a small wrapper file is created on the fly. This is necessary if we
want to be using miniaudio's high level API (without rewriting all the
interfaces for a custom decoder.)